Improvement of docker file
I added a staging process in the build process for the image.

In the final docker image there will be installed hunspell to support the spell checking.

This is just quick try. We have to look on the user. Maybe we have to change the user at the end, so that the server ist not running as root.